feat: add robust ship rotation functionality
- Add robust ship rotation system that satisfies modern, gaming user
experience standards (i.e., you can rotate the ship on a 360 degree axis
in increments of 90 degrees regardless of the position of other ships at
rotation time. It ensures the ship remains in bounds as well as avoids
collisions with other ships. Wouldn’t want them sinking now ;).
- Re-added `fleetBuilder` as a dependency in  `BattleShipBoardController`
instantiation. The rotation system within the controller requires access
to all ships in the fleet. When any ship is placed/replaced, the
repository (class) upserts
`_inBoundRotationalPlacePieceParamsForFleet`… which stores `PlacePiece`
arguments for rotational positions for all placed ships that are within
the bounds of the board (it does not account for potential collisions).
The ships can be rotated on an axis of 360 degrees in increments of
90 degrees. The repository stores `PlacePiece` arguments that reposition
the ship to one of the 3 other rotated positions (i.e,, 90deg, 180deg,
and 270deg if in non-rotated position).
- Renamed `ValidRotationalPositionMap` to be `RotationalPositionMap`
because its value has been updated to include invalid
rotational positions. The ship rotation logic is no longer dependent on
the rotational position map that had only valid positions; However, it
is still used in the rotational position validation process.
- Abstracted `isRotatedPositionValid` out of the `rotatePiece` method
(now, a private helper method of the class) as a design choice. I
decided that it deserved its own code block considering it contains
jarring logic-intensive operations. It being a private method tells the
reader, “this is intentionally separated because it is not your average
state change”.
- Renamed key in `BattleshipBuilder` property
`rotationalPivotConfigurations` from `currentAngleOfRotation` to
`transientAngleOfRotation` to imply that the value is not permanent,
as its value is mutated through background processes.
- Enabled ship colors in `SASS` to enhance the MVP UI.
- Removed unused `getValidShipPositions`,
`battleshipBoardController` parameter.
- Formally Updated types: `FleetValidRotationalParams` to be
`InBoundRotationalPlacePieceParamsForFleet` and `currentAngleOfRotation`
to be `transientAngleOfRotation`.

💭