refactor(app): switch to class for better lifecycle management
- Refactors `createDOM` to be `DOMController` class and now handles the
instructions light-box instantiation and controller attachment. This
controller is also injected into the `Main` class constructor where its
`openLightbox` method is set as the callback for the instruction
buttons' on-click event listener.
- Refactors `createMain` function to be `MainComponent` class where the
`instructionsLightboxController` is accepted as an argument in its
constructor. This enables a streamlined method of integrating the
`CycloneLightboxController` dialog handler.
- Also, the `MainComponent` class exposes the three main containers that
comprise it. This allows seamless, simultaneous content swapping in each
of them from outside the class.
- Removes `instructions-dialog.ts` because `DOMController` now handles
instantiation and handling.